==> 2b - TVMaze API.ipynb <==
In this notebook, we'll be using requests and some custom functions to take the list of shows we got in the previous notebook and search for them in the TVMaze API, returning some additional metadata and the IMDB ID of each show.

import pandas as pd
import numpy as np
import requests
import time
import re
shows = pd.read_pickle("ismyshowcancelled_final.pkl")

# Building a requester for the shows we want more information on
def show_request(show):
    
    # proper formatting
    show = '-'.join(show.lower().split()).replace(':','')
    
    # define params for json request
    params = {'q':show}#,'embed':'episodes'}

    # send request via requests
    r = requests.get('http://api.tvmaze.com/singlesearch/shows?',params=params)
    
    return r

# Extract the information we want for each show
def extract_info(r):
    
    # Check that the request worked
    if r.reason != 'OK':
        print('No Matches found for {}'.format(show))
        attributes = 7*[np.nan]
    
    else:
        data = r.json()

        # Pulling the show info we want
        name = data['name']
        tv_id = data['id']
        imdb = data['externals']['imdb']
        prem_date = data['premiered']
        rating = data['rating']['average']
        runtime = data['runtime']
        if len(data['schedule']['days']) == 1: 
            ep_day = data['schedule']['days'][0]
        else: ep_day = 'Multiple'

        attributes = [name,tv_id,imdb,prem_date,rating,runtime,ep_day]

        return attributes

# Create a dataframe to house this data
tvmaze = pd.DataFrame(columns=['name','tv_id','imdb','prem_date','rating','runtime','ep_day'])
index = 0

# Creating a loop to hopefully get all this show info!
for show in shows['title']:
    
    # Make the request for the show
    r = show_request(show)
    
    # Extract the attributes
    attributes = extract_info(r)
    
    # Add these as an entry in the dataframe (nulls populate if our search wasn't successful)
    tvmaze.loc[index] = attributes
    
    index += 1
    
    # Sleep for half a second so we don't get "too many requests" errors
    time.sleep(0.5)

tvmaze.to_csv("tvmaze_tmp_1.csv")

tvmaze.head()

tvmaze = pd.read_csv("tvmaze_tmp_1.csv",index_col=0)

# Shows we need to redo
re_search_shows = ['24: Live Another Day',
 'Extreme Makeover: Weight Loss Edition',
 'Lopez Tonight',
 'Match Game (2016)',
 'Numb3rs (aka Numbers)',
 'Panic 911',
 'Rock Center with Brian Williams',
 'Teach: Tony Danza',
 'The $100,000 Pyramid (2016)',
 'The Gong Show (2017)',
 'The Martha Stewart Show',
 'The X-Files (2016)',
 'To Tell The Truth (2016)',
 'Twin Peaks (2017)']

# Their indices in our original data
indices = []
for item in re_search_shows:
    indices.append(shows[shows['title'] == item].index[0])

# Redoing searches with some simple rules to see if it helps
new_terms = []
for show in re_search_shows:
    show = re.sub(r'\(.*\)','', show).strip()
    new_terms.append(show)
    
# Changing the pyramid one
new_terms[8] = 'the-100000-pyramid'

new_terms

This would probably work better with the above terms. A more surefire way, though, would be to just look up the IMDB IDs for these ten and index off those, just to be safe. This has the added benefit that even if we don't find them in tvmaze, we'll still be able to look up their IMDB information later.

imdb_ids= ['tt1598754',
 'tt1713288',
 'tt1489432',
 'tt5672484',
 'tt0433309',
 'tt2738058',
 'tt2084611',
 'tt1442170',
 'tt5330088',
 'tt6128376',
 'tt3450386',
 'tt0106179',
 'tt5817158',
 'tt4093826']

imdb_dict = {x:y for x,y in zip(imdb_ids,indices)}

# Building a requester for the shows we want more information on
def imdb_request(imdb):
    
    # define params for json request
    params = {'imdb':imdb}

    # send request via requests
    r = requests.get('http://api.tvmaze.com/lookup/shows?',params=params)
    
    return r

# Extract the information we want for each show
def extract_imdb_info(r,imdb):
    
    # Check that the request worked
    if r.reason != 'OK':
        print('No Matches found for {}'.format(shows.iloc[imdb_dict[imdb_id],:].loc['title']))
        attributes = np.array([np.nan,np.nan,imdb,np.nan,np.nan,np.nan,np.nan])
        attributes[2] = imdb
    
    else:
        data = r.json()

        # Pulling the show info we want
        name = data['name']
        tv_id = data['id']
        imdb = data['externals']['imdb']
        prem_date = data['premiered']
        rating = data['rating']['average']
        runtime = data['runtime']
        if len(data['schedule']['days']) == 1: 
            ep_day = data['schedule']['days'][0]
        else: ep_day = 'Multiple'

        attributes = [name,tv_id,imdb,prem_date,rating,runtime,ep_day]
        print(attributes)
        return attributes

redo_shows = pd.DataFrame(data={'index':indices,'re_search_shows':re_search_shows,'imdb':imdb_ids})
redo_shows

# Let's try to ping for this info!
for imdb_id, index in zip(redo_shows['imdb'],redo_shows['index']):
    
    # Request based on the imdb id
    r = imdb_request(imdb_id)
    
    # Extract from the result, saving the id if we don't get anything back
    attributes = extract_imdb_info(r,imdb_id)        
    
    tvmaze.iloc[index,:] = attributes

tvmaze[(tvmaze['name'].isnull()) & (tvmaze['imdb'].isnull())]

Don't quite understand why this isn't at least appending the IMDB IDs. Doing that here...

for imdb,index in zip(redo_shows['imdb'],redo_shows['index']):
    print(index,imdb)

for imdb,index in zip(redo_shows.copy()['imdb'],redo_shows.copy()['index']):
    tvmaze.loc[:,'imdb'].iloc[index] = imdb

tvmaze.loc[:,'imdb'].iloc[924]

tvmaze[(tvmaze['name'].isnull()) & (tvmaze['imdb'].isnull())]

tvmaze.to_csv("tvmaze_tmp_2.csv")

Phew, okay now we at least have an IMDB ID or tvmaze supplemental information for each show. Let's do one last pass with the cleaned show names we abandoned earlier.

new_show_dict = {x:y for x,y in zip(new_terms,indices)}
for item in ['Panic 911','Extreme Makeover: Weight Loss Edition','Numb3rs','the-100000-pyramid','The X-Files']:
    new_show_dict.pop(item)
new_show_dict

We have values for Extreme Weight Loss, Numb3rs, Panic 911, 100k pyramid, and x files, so we won't iterate over those again

tvmaze.loc[:, tvmaze.columns != 'imdb'].iloc[924]

# Creating a loop to hopefully get all this show info!
for show,index in new_show_dict.items():
    
    # Make the request for the show
    r = show_request(show)
    
    # Extract the attributes
    attributes = extract_info(r)
    
    # Making sure IMDB id is not overwritten
    if attributes is not None:
        #attributes = [x for i,x in enumerate(attributes) if i!=2]
        print(attributes)

Cool, four more matches. I'll do this manually since I'm having trouble not overriding the preexisting IMDB IDs that we've independently collected and verified.

redo_shows

tvmaze.iloc[805] = ['Match Game', 16694, 'tt5672484', '2016-06-26', None, 60, 'Thursday']
tvmaze.iloc[1289] = ['The Gong Show', 21746, 'tt6128376', '2017-06-22', 5, 60, 'Thursday']
tvmaze.iloc[1451] = ['To Tell the Truth', 16041, 'tt5817158', '2016-06-14', None, 60, 'Monday']
# Manually changing this to just go with the newer version of Twin Peaks
tvmaze.iloc[1482] = ['Twin Peaks', 156, 'tt4093826', '1990-04-08', 8.2, 60, 'Sunday']

tvmaze[tvmaze['name'].isnull()]

Pretty good. Only five total shows unaccounted for in the tvmaze data, and for all of those we have the IMDB IDs to pull more data.

tvmaze.to_csv("tvmaze_tmp_3.csv")

tvmaze.iloc[737]

shows.iloc[737]

import pandas as pd

tvmaze = pd.read_csv("tvmaze_tmp_3.csv",index_col=0)

tvmaze.dtypes

# Doing some downcasting to save to pkl...
tvmaze[['name','imdb','tv_id','ep_day']] = tvmaze[['name','imdb','tv_id','ep_day']].apply(lambda x: x.astype('category'))
