==> 2c - OMDB Initial Supplementation.ipynb <==
Before we do a full pull from OMDB, we'll need to ping the OMDB API for any shows for which we couldn't pick up an IMDB ID from TVMaze (which generally has better search matching than OMDB). To do this, we need to be very careful about timing. I've written a custom wrapper and requests function that sleeps if a request times out due to server overload, such that the dataset does not need to be gathered piecemeal.

import pandas as pd
import numpy as np
import requests
import time
import re
from functools import wraps
import yaml
shows = pd.read_pickle("ismyshowcancelled_final.pkl")
tvmaze = pd.read_csv("tvmaze_tmp_3.csv",index_col=0)
auth = yaml.load(open('omdb.yaml'))
key = auth['OMDB']['Key']

There's about ~170 shows that don't have an IMDB ID according to tvmaze. I suspect, though, that this is a shortcoming in their database and not an actual indication that IMDB has no record of these shows. Let's first get a list of the ones we want to search for.

# This dataframe is going to house the shows with missing IMDB IDs, 
# as well as the IDs that we eventually (hopefully) will return
missing_imdb = pd.DataFrame(tvmaze[tvmaze['imdb'].isnull()]['name'])
missing_imdb['found_id'] = np.nan
missing_imdb['found_name'] = np.nan
missing_imdb.head()

# A wrapper to retry the request function many times, with a sleep in between
def retry_multi(max_retries):
    """ Retry a function `max_retries` times. """
    def retry(func):
        @wraps(func)
        def wrapper(*args, **kwargs):
            num_retries = 0 
            while num_retries <= max_retries:
                try:
                    ret = func(*args, **kwargs)
                    break
                except:
                    if num_retries == max_retries:
                        raise
                    num_retries += 1
                    time.sleep(15)
            return ret 
        return wrapper
    return retry

# Building a requester for the IMDB IDs from OMDB (note that this differs from our tvmaze equivalent function)
@retry_multi(10)
def request_info(show):
    
    # proper formatting
    show = '-'.join(show.lower().split()).replace(':','')
    
    # define params for json request
    params = {'apikey':key,'t':show,'type':'series'}

    # send request via requests
    response = requests.get('http://www.omdbapi.com/?',params=params)
    
    return response

# Building functions to extract just the IMDB IDs and names (just to cross-check that the result wasn't erroneous)
# Later, we'll return to scrape all the information we want once we have a dataframe filled with necessary IDs
def get_json(response):
    
    # Just return the json, if the search was successful
    if response.json()['Response'] == 'True':
        success = True
    elif response.json()['Response'] == 'False':
        success = False
        
    # Return a dict indicating success/failure of request, with json
    return {'json':response.json(), 'success':success}

def extract_imdb(json_response):
    
    # get the id, if successful pull
    if json_response['success'] == True:
        return(json_response['json']['imdbID'])

def extract_title(json_response):
    
    # get the name, if successful pull. this is just for cross-checking and we'll delete it later
    if json_response['success'] == True:
        return(json_response['json']['Title'])

# Now writing a loop to ping for requests!
for show,index in zip(missing_imdb.name,missing_imdb.index):
    
    # Request the show from OMDB
    r = request_info(show)
    json_response = get_json(r)
    
    # Getting imdb id and title (will not collect anything if unsuccessful json response)
    imdb = extract_imdb(json_response)
    title = extract_title(json_response)
    
    # Doing a double check (also in function) that we got a good response, then inputting info back into our df
    if json_response['success'] is True:
        
        # Letting us know what happened
        print('Match for {}: {}. ID is {}'.format(show,title,imdb))
        
        # Adding this result back to our dataframe
        missing_imdb.loc[missing_imdb.index == index,'found_name'] = title
        missing_imdb.loc[missing_imdb.index == index,'found_id'] = imdb
        #missing_imdb.iloc[index].loc['found_id'] = imdb
        #missing_imdb.iloc[index].loc['found_name'] = title
        
    else:
        print('No match found for {}'.format(show))
    
    time.sleep(2)

missing_imdb.tail(30)

Success! We got a good number of matches, and the leftovers are few enough that I can do a quick check for the remainder and just write those into the CSV. Additionally, those matches we did get seem entirely accurate.

For now, we'll add these IMDB IDs back into the original dataframe nad go from there!

for show in missing_imdb['name']:
    tvmaze.loc[tvmaze['name'] == show,'imdb'] = missing_imdb.loc[missing_imdb['name'] == show,'found_id']

Doing some spot checking...

tvmaze.loc[tvmaze['name'] == 'The Hasselhoffs','imdb']

tvmaze.loc[tvmaze['name'] == 'The Outpost']

tvmaze.isnull().sum()

tvmaze[tvmaze['imdb'].isnull()]

Awesome, we've greatly reduced the number of missing IDs. Let's save this to CSV and then I'll fill in the remainder myself.

tvmaze.to_csv("tvmaze_tmp_4_more_ids.csv")

I found matches for all but 2: Ink Master Redemption (which is considered to be the same thing as Ink Master by IDMB, which we already have in our dataset), and Crash Course Engineering (this is a YouTube series). We'll probably delete both once we get into the analysis and modeling phases.

tvmaze[tvmaze['tvmaze_name']=='Caprica']

tvmaze = pd.read_csv("tvmaze_tmp_5_more_ids_MORE ADDITIONS.csv",index_col=0)

tvmaze.isnull().sum()

Awesome! Only one thing remains for now. Let's merge this with our shows df. Luckily, they rely on the exact same indices.

l = []
for col in tvmaze.columns:
    l.append('tvmaze_'+col)

l

l[2] = 'imdb_id'

l

tvmaze.columns = l

tvmaze.head()

tvmaze.isnull().sum()

tvmaze.iloc[234:]

shows = shows.join(tvmaze)

shows.head()

Saving to pkl and csv...

shows.to_csv("full_shows_tmp_1_tvmaze_plus_imdbids.csv")

shows.to_pickle("full_shows_tmp_1_tvmaze_plus_imdbids.pkl")

shows.shape

shows.isnull().sum()
