==> 2d - Full OMDB Scrape.ipynb <==
Now that we've got IMDB IDs for nearly every show in our original scrape, let's get some additional info from IMDB that TVmaze couldn't provide. This includes stuff like actors, writers, and the IMDB score, which is likely more meaningful than the TVMaze result given the former's larger user base. We'll again need to use the custom wrapper and requests function to avoid overloading the server and breaking the pull.

import pandas as pd
import numpy as np
import requests
import time
import re
import yaml
import scrapy
import pickle
from functools import wraps
shows = pd.read_pickle("full_shows_tmp_1_tvmaze_plus_imdbids.pkl")
auth = yaml.load(open('omdb.yaml'))
key = auth['OMDB']['Key']

# A wrapper to retry the request function many times, with a sleep in between
def retry_multi(max_retries):
    """ Retry a function `max_retries` times. """
    def retry(func):
        @wraps(func)
        def wrapper(*args, **kwargs):
            num_retries = 0 
            while num_retries <= max_retries:
                try:
                    ret = func(*args, **kwargs)
                    break
                except:
                    if num_retries == max_retries:
                        raise
                    num_retries += 1
                    time.sleep(15)
            return ret 
        return wrapper
    return retry

# Bringing in our requester from the previous step (OMDB Initial Supplementation)
# But with the addition of functionality from ratelimit
#@sleep_and_retry
#@limits(calls=10,period=10)
@retry_multi(10)
def request_info(imdb_id):
    
    # define params for json request
    params = {'apikey':key,'i':imdb_id,'plot':'full'}

    # send request via requests
    try:
        response = requests.get('http://www.omdbapi.com/?',params=params,timeout=10)
        time.sleep(0.25)
    except:
        while num_tries <= 15:
            num_tries = 1
            response = requests.get('http://www.omdbapi.com/?',params=params,timeout=30)
            num_tries += 1
            time.sleep(1)
    
    if response.status_code != 200:
        raise Exception('API response: {}'.format(response.status_code))
    
    return response

def get_json(response):
    
    # Just return the json, if the search was successful
    if response.json()['Response'] == 'True':
        success = True
    elif response.json()['Response'] == 'False':
        success = False
        
    # Return a dict indicating success/failure of request, with json
    return {'json':response.json(),'success':success}

Looping through our results...

# This will be a list of the JSONs we receive
feed = []

for show in shows['imdb_id']:
    r = request_info(show)
    json = get_json(r)
    if json['success'] == True:
        print('Success: {}'.format(json['json']['Title']))
        feed.append(json['json'])
    else:
        print('Request worked, but show not found.')

len(feed)

Hooray! Let's save the list for safekeeping

print('Checkpoint: {} items'.format(len(feed)))
with open('feed_tmp_3.pkl', 'wb') as f:
    pickle.dump(feed, f)

Now let's go ahead and try to wrangle this data into a dataframe!

imdb_data = pd.DataFrame(feed)
imdb_data.head()

# Renaming columns...
newcols = []
for col in imdb_data.columns:
    if 'imdb' in col:
        col = col.strip('imdb')
    col = 'imdb_'+col
    col = col.lower()
    newcols.append(col)
    
imdb_data.columns = newcols
print(imdb_data.columns)

shows.columns

shows = shows.merge(imdb_data,how='left',on='imdb_id')

shows.columns

shows[shows['imdb_id'].isnull()]

Above are the two that didn't have good matches on IMDB (as discussed on OMDB). They still seemed to have been merged fine.

Let's checkpoint to pickle...

shows.to_pickle("full_shows_tmp_2_tvmaze+imdb.pkl")

shows = pd.read_pickle("full_shows_tmp_2_tvmaze+imdb.pkl")

len(shows)

shows['title'].value_counts()

Looks like we did have a couple weird instances where a show showed up multiple times. Let's clear those up.

cleaner = shows.copy()
cleaner['ones'] = 1
shows['title_cumsum'] = cleaner.groupby('title')['ones'].transform(np.cumsum)
shows['title_dupes'] = cleaner.groupby('title')['ones'].transform('count')
shows['imdbid_dupes'] = cleaner.groupby('title')['ones'].transform('count')
shows['imdbid_cumsum'] = cleaner.groupby('title')['ones'].transform(np.cumsum)

shows = shows[(shows['imdbid_cumsum'] == 1) | (shows['imdbid_cumsum'].isnull())]

shows.shape

Looks like we're back to the right number. Pickling again - next, data cleaning & exploration!

shows = shows.drop(['title_cumsum','title_dupes','imdbid_dupes','imdbid_cumsum'],axis=1)
shows.columns

shows.to_pickle("full_shows_tmp_6_tvmaze+imdb.pkl")

shows = pd.read_pickle("full_shows_tmp_6_tvmaze+imdb.pkl")
