==> 8b - word2vec Embedding with Keras.ipynb <==
Many thanks to Jason Brownlee's [blog post](https://machinelearningmastery.com/use-word-embedding-layers-deep-learning-keras/) on this topic, which was extremely helpful in getting this up and running considering I have not had a chance to do the Keras specialization yet.

import keras
import pandas as pd
import numpy as np
from sklearn.externals import joblib
from sklearn import metrics
shows = pd.read_pickle("no_na_pre2017_v4.pkl")

import gensim
model = gensim.models.KeyedVectors.load_word2vec_format('/Users/perrypetra-wong/Dropbox/Thinkful/Lessons/Capstone/GoogleNews-vectors-negative300.bin', binary=True)
#model = gensim.models.KeyedVectors.load_word2vec_format('GoogleNews-vectors-negative300.bin', binary=True)

For our text documents, let's start with a concatenation of the three text columns, and define our labels: cancellation.

docs = shows['tagline'].astype('O') + shows['synopsis'].astype('O')
docs[0]

labels = shows['two_season_cancel']

Limiting our word vectors to just those present in the text.

import string

embeddings_index = {}

for wordstring in docs:
    
    if isinstance(wordstring,float):
        pass
    else:
        # In case of null
        punct_remover = str.maketrans('', '', string.punctuation)

        # Removing punctuation and splitting apart words
        punct_stripped = wordstring.translate(punct_remover)
        wordlist = punct_stripped.split()

        # If word is in Google model, add it its corresponding vec to our running
        # subset dict
        for word in wordlist:
            word = word.lower()
            if word in model:
                if word not in embeddings_index:
                    embeddings_index[word] = model[word]
                    
del model

#joblib.dump(embeddings_index,"Extras/embeddings_index.pkl")
embeddings_index = joblib.load("Extras/embeddings_index.pkl")

# Tokenize the texts to save mappings from words to integers
t = keras.preprocessing.text.Tokenizer()
t.fit_on_texts(docs)
vocab_size = len(t.word_index) + 1

encoded_docs = t.texts_to_sequences(docs)

print(encoded_docs[0:1])

# We need to pad the length of these to a fixed amount that won't be exceeded
max_length = 25
padded_docs = keras.preprocessing.sequence.pad_sequences(encoded_docs, maxlen=max_length, padding='post')

Next we need to create a matrix of one embedding per word in the training set. We can do that by enumerating all unique words in the Tokenizer.word_index and locating the corresponding vector in the word2vec subset we've created above.

# Weight matrix for each word in training doc
embedding_matrix = np.zeros((vocab_size, 300))
for word, i in t.word_index.items():
    embedding_vector = embeddings_index.get(word)
    if embedding_vector is not None:
        embedding_matrix[i] = embedding_vector

e = keras.layers.Embedding(vocab_size,300,weights=[embedding_matrix],input_length=max_length,trainable=False)

# Define the model
model = keras.Sequential()
model.add(e)
model.add(keras.layers.Flatten())
model.add(keras.layers.Dense(1,activation='sigmoid'))

# Compile the model
model.compile(optimizer='adam', loss='binary_crossentropy', metrics=['acc'])

print(model.summary())

model.fit(padded_docs,labels,epochs=50,verbose=1)

y_pred = model.predict(padded_docs)

metrics.accuracy_score(labels,np.round(y_pred))

Obviously it performs quite well on the training set, but this was more to just check that the model worked. Let's now run it in a cross-validated fashion in the same way that we've done for our other models.

from sklearn.model_selection import KFold
kf = KFold(n_splits=3)

# reindexing docs and labels to avoid any errors there
docs.index = np.arange(0,len(docs))
docs = docs.reindex()
labels.index = (np.arange(0,len(labels)))

for train,test in kf.split(docs):
      
    # Train and test
    docs_train = docs[train]
    labels_train = labels[train]
    docs_test = docs[test]
    labels_test = labels[test]
    
    # Tokenize the texts to save mappings from words to integers
    t = keras.preprocessing.text.Tokenizer()
    t.fit_on_texts(docs_train)
    vocab_size = len(t.word_index) + 1

    encoded_docs = t.texts_to_sequences(docs_train)
    
    # Capping the lengths
    max_length = 25
    padded_docs = keras.preprocessing.sequence.pad_sequences(encoded_docs, maxlen=max_length, padding='post')
    
    # Weight matrix for each word in training doc
    embedding_matrix = np.zeros((vocab_size, 300))
    for word, i in t.word_index.items():
        embedding_vector = embeddings_index.get(word)
        if embedding_vector is not None:
            embedding_matrix[i] = embedding_vector
    
    e = keras.layers.Embedding(vocab_size,300,weights=[embedding_matrix],input_length=max_length,trainable=False)
    
    # Define the model
    model = keras.Sequential()
    model.add(e)
    model.add(keras.layers.SimpleRNN(10, activation='relu', input_shape=(300,), unroll=True, return_sequences=False))
    #model.add(keras.layers.Flatten())
    model.add(keras.layers.Dense(1,activation='sigmoid'))

    # Compile the model
    model.compile(optimizer='adam', loss='binary_crossentropy', metrics=['acc'])
    
    model.fit(padded_docs,labels_train,epochs=50,verbose=0)
    
    # Test set
    encoded_test = t.texts_to_sequences(docs_test)
    padded_docs_test = keras.preprocessing.sequence.pad_sequences(encoded_test, maxlen=max_length, padding='post')
    
    y_pred = model.predict(padded_docs_test)
    loss,accuracy = model.evaluate(padded_docs_test,labels_test)
    
    print('Accuracy on first fold: {0:.2%}'.format(accuracy))

# Confirming how the accuracy seems to be calculated
metrics.accuracy_score(np.round(y_pred),labels_test)

Sadly, we're still looking at pretty poor results. Even though our neural net can easily converge on 100% training accuracy within 50 epochs, accuracy falls apart when applied to holdouts. As such, I don't think adding layers to our net and optimizing would be very fruitful, since we seem to be only hurting our predictive value when we attempt more complex learning from the training documents. In other words, we're observing the bias-variance tradeoff.

This is not to say that NLP doesn't have some cool applications in this dataset. Out of curiosity, let's see how this network does when assigned to something else, like telling whether or not a show is a reality show!

labels = shows['Reality']

from sklearn.model_selection import KFold
kf = KFold(n_splits=3)

# reindexing docs and labels to avoid any errors there
docs.index = np.arange(0,len(docs))
docs = docs.reindex()
labels.index = (np.arange(0,len(labels)))

for train,test in kf.split(docs):
      
    # Train and test
    docs_train = docs[train]
    labels_train = labels[train]
    docs_test = docs[test]
    labels_test = labels[test]
    
    # Tokenize the texts to save mappings from words to integers
    t = keras.preprocessing.text.Tokenizer()
    t.fit_on_texts(docs_train)
    vocab_size = len(t.word_index) + 1

    encoded_docs = t.texts_to_sequences(docs_train)
    
    # Capping the lengths
    max_length = 35
    padded_docs = keras.preprocessing.sequence.pad_sequences(encoded_docs, maxlen=max_length, padding='post')
    
    # Weight matrix for each word in training doc
    embedding_matrix = np.zeros((vocab_size, 300))
    for word, i in t.word_index.items():
        embedding_vector = embeddings_index.get(word)
        if embedding_vector is not None:
            embedding_matrix[i] = embedding_vector
    
    e = keras.layers.Embedding(vocab_size,300,weights=[embedding_matrix],input_length=max_length,trainable=False)
    
    # Define the model
    model = keras.Sequential()
    model.add(e)
    model.add(keras.layers.Flatten())
    model.add(keras.layers.Dense(30,activation='relu'))
    model.add(keras.layers.Dense(1,activation='sigmoid'))


    # Compile the model
    model.compile(optimizer='adam', loss='binary_crossentropy', metrics=['acc'])
    
    model.fit(padded_docs,labels_train,epochs=50,verbose=0)
    
    # Test set
    encoded_test = t.texts_to_sequences(docs_test)
    padded_docs_test = keras.preprocessing.sequence.pad_sequences(encoded_test, maxlen=max_length, padding='post')
    
    y_pred = model.predict_classes(padded_docs_test)
    loss,accuracy = model.evaluate(padded_docs_test,labels_test)
    
    print('Accuracy on first fold: {0:.2%}'.format(accuracy))

shows['primary_genre'].value_counts().apply(lambda x:x/len(shows))

shows['Reality'].mean()

Pretty impressive for a quick solution, with 80% accuracy (even if only 5% above baseline). Let's see how it does with all genres.

from sklearn.preprocessing import LabelEncoder
encoder = LabelEncoder()

# Need to properly encode the outputs
labels = pd.Series(encoder.fit_transform(shows['primary_genre']))
encoded_labels = encoder.fit_transform(shows['primary_genre'])
one_hot_labels = keras.utils.np_utils.to_categorical(encoded_labels)
one_hot_labels = np.array(pd.get_dummies(encoded_labels))

from sklearn.model_selection import KFold
kf = KFold(n_splits=3)

# reindexing docs and labels to avoid any errors there
docs.index = np.arange(0,len(docs))
docs = docs.reindex()

for train,test in kf.split(docs):
      
    # Train and test
    docs_train = docs[train]
    labels_train = one_hot_labels[train]
    docs_test = docs[test]
    labels_test = one_hot_labels[test]
    
    # Tokenize the texts to save mappings from words to integers
    t = keras.preprocessing.text.Tokenizer()
    t.fit_on_texts(docs_train)
    vocab_size = len(t.word_index) + 1

    encoded_docs = t.texts_to_sequences(docs_train)
    
    # Capping the lengths
    max_length = 40
    padded_docs = keras.preprocessing.sequence.pad_sequences(encoded_docs, maxlen=max_length, padding='post')
    
    # Weight matrix for each word in training doc
    embedding_matrix = np.zeros((vocab_size, 300))
    for word, i in t.word_index.items():
        embedding_vector = embeddings_index.get(word)
        if embedding_vector is not None:
            embedding_matrix[i] = embedding_vector
    
    e = keras.layers.Embedding(vocab_size,300,weights=[embedding_matrix],input_length=max_length,trainable=False)
    
    # Define the model
    model = keras.Sequential()
    model.add(e)
    model.add(keras.layers.Flatten())
    model.add(keras.layers.Dense(30))
    model.add(keras.layers.Dense(30))
    model.add(keras.layers.Dense(6,activation='softmax'))

    # Compile the model
    model.compile(optimizer='adam', loss='categorical_crossentropy', metrics=['acc'])
    
    model.fit(padded_docs,labels_train,epochs=50,verbose=0)
    
    # Test set
    encoded_test = t.texts_to_sequences(docs_test)
    padded_docs_test = keras.preprocessing.sequence.pad_sequences(encoded_test, maxlen=max_length, padding='post')
    
    y_pred = model.predict_classes(padded_docs_test)
    loss,accuracy = model.evaluate(padded_docs_test,labels_test)
    
    print('Accuracy on first fold: {0:.2%}'.format(accuracy))

Very cool. Almost a 2/3 accuracy vs. a baseline guess of the dominant class, which would have an accuracy of ~40%. Again, this doesn't help our cause of predicting cancellation, which seems like it's just not solvable with NLP given the current information we have.

model.summary()
